Remove Monitors class and avoid duplicate state
The monitors class is no longer needed, because it had different
responsibilities in widely different parts of the application.
Therefore, it has been removed. We no longer have any duplicate monitor
state - instead the last state is retreived from FLTK and then the
widget uses native widget to store button state.

The storage that was previously handled by the Monitors class is now
instead handled by a custom StringParameter subclass
(MonitorIndicesParameter). By using its getters and setters, the FLTK
indices for the external representation can be retrieved to be used in
the application