Fix the position of the F8 overlay
After introducing fullscreen on selected monitors, the F8 overlay broke
because it always tried to use the primary monitor. Furthermore, the
position of the overlay was sometimes wrong, because the coordinates
were not relative to the actual viewport.

Now, the overlay will be shown on the monitor having the smallest FLTK
index that is enclosed by the viewport. The coordinates are adjusted
such that the overlay is centered correctly too.