Refactor get_monitor_name to improve compability
The way we fetched the monitor name on macOS were not compatible with
older versions of the operating system. It has been replaced with
another approach. Doing this, some changes were made to how the display
name is fetched on all platforms, in order to avoid relying on memory
still being allocated, that in reality, might not be.